==> notebooks/08_xgboost_model.ipynb <==
import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
from src.config import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "transformed_features_and_target_top50.parquet")
df

from datetime import datetime

from src.data_utils import split_time_series_data

X_train, y_train, X_test, y_test = split_time_series_data(
    df,
    cutoff_date=datetime(2025, 7, 1, 0, 0, 0),
    target_column="target"
)

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

X_train

past_ride_columns = [c for c in X_train.columns if c.startswith("rides_")]
X_train_only_numeric = X_train[past_ride_columns]
X_test_only_numeric = X_test[past_ride_columns]

#!pip install xgboost

# Cell 1: Uninstall current xgboost
#!pip uninstall xgboost -y

# Cell 2: Install xgboost via conda (recommended)
#!conda install -c conda-forge xgboost -y

# Cell 3: Restart the kernel to apply changes
#import IPython
#IPython.Application.instance().kernel.do_shutdown(True)

import xgboost as xgb
model = xgb.XGBRegressor(max_depth=10)
model.fit(X_train_only_numeric, y_train)

from sklearn.metrics import mean_absolute_error
predictions = model.predict(X_test_only_numeric)

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae:.4f}")
