==> notebooks/10_ARMA_model.ipynb <==
# run in a notebook cell
%pip install statsmodels

# run in a notebook cell so it installs into the current kernel
%pip install prophet

import sys
import os
import pandas as pd
import numpy as np
from datetime import datetime
from scipy.fft import fft
from statsmodels.tsa.arima.model import ARIMA
from prophet import Prophet
from sklearn.metrics import mean_absolute_error
from dotenv import load_dotenv
# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd
from src.config import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "transformed_features_and_target_top50.parquet")
df

from datetime import datetime

from src.data_utils import split_time_series_data

X_train, y_train, X_test, y_test = split_time_series_data(
    df,
    cutoff_date=datetime(2025, 7, 1, 0, 0, 0),
    target_column="target"
)

print(X_train.shape)
print(y_train.shape)
print(X_test.shape)
print(y_test.shape)

def add_fft_features(X):
    fft_features = []
    for col in X.columns:
        if col.startswith('rides_'):
            fft_result = fft(X[col])
            fft_features.append(pd.Series(np.abs(fft_result), name=f'fft_{col}'))
    return pd.concat([X] + fft_features, axis=1)

X_train_fft = add_fft_features(X_train)
X_test_fft = add_fft_features(X_test)

def fit_arima(train, test, order=(5,1,0)): #order=(1,1,1)
    # Fit ARIMA model
    model = ARIMA(train, order=order)
    results = model.fit()

    # Generate forecasts
    forecast = results.forecast(steps=len(test))
    return forecast

arima_forecast = fit_arima(y_train, y_test)
arima_mae = mean_absolute_error(y_test, arima_forecast)
print(f"ARIMA MAE: {arima_mae:.4f}")
